+++ src/main/java/com/example/demo/service/UserService.java
@@ -6,6 +6,7 @@ import com.example.demo.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -17,17 +18,26 @@ public class UserService {
     // Create
     public User save(AddUserRequest request) {
 
-        Optional<User> registeredUserId = userRepository.findById(request.getUserId());
+        return (User) userRepository.findById(request.getUserId())
+                .map(user -> {
+                    throw new IllegalArgumentException("User already exists");
+                }).orElseGet(() -> userRepository.save(request.toEntity()));
 
-        if(registeredUserId.isEmpty()) {
-            return userRepository.save(request.toEntity());
-        } else {
-            throw new IllegalArgumentException("User already exists");
-        }
+
+//        Optional<User> registeredUserId = userRepository.findById(request.getUserId());
+//
+//        if(registeredUserId.isEmpty()) {
+//            return userRepository.save(request.toEntity());
+//        } else {
+//            throw new IllegalArgumentException("User already exists");
+//        }
     }
 
     // Delete
     public void delete(String userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NoSuchElementException("User does not exist");
+        }
         userRepository.deleteById(userId);
     }
 }
